+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 
 ### Changed
 * Show a better error message when the folder does not exist. [#2341](https://github.com/fsprojects/fantomas/issues/2341)
+* Update FCS to 'Use SynLongIdent in SynOpenDeclTarget.ModuleOrNamespace', commit c18989124a92cc26a736d89137254f1ca84d4880
+
+### Fixed
+* global keyword is escaped with single-backticks when used as part of a namespace/module open. [#2274](https://github.com/fsprojects/fantomas/issues/2274)
+* global namespace should not be escaped. [#2366](https://github.com/fsprojects/fantomas/issues/2366)
 
 ## [5.0.0-beta-002] - 2022-07-19